Add backward compatible implementation for redirect_stdout and redirect_stderr
also require enum34 for Python 3.3
